(#16) Enable uninstalling Go
Adds `$ensure` parameter that may be set to `absent` to remove the Go
installation.

+++ manifests/init.pp
@@ -4,6 +4,9 @@
 #
 # Most people will not need to change any parameter other than `$version`.
 #
+# @param ensure
+#   * `present`: Make sure go is installed.
+#   * `absent`: Make sure go is uninstalled.
 # @param version
 #   The version of Go to install. You can find the latest version number at
 #   https://go.dev/dl/
@@ -18,15 +21,16 @@
 # @param source
 #   URL to actual archive.
 class golang (
-  String[1]        $version       = '1.19.1',
-  Array[String[1]] $link_binaries = ['go', 'gofmt'],
-  String[1]        $source_prefix = 'https://go.dev/dl',
-  String[1]        $os            = $facts['kernel'] ? {
+  Enum[present, absent] $ensure        = present,
+  String[1]             $version       = '1.19.1',
+  Array[String[1]]      $link_binaries = ['go', 'gofmt'],
+  String[1]             $source_prefix = 'https://go.dev/dl',
+  String[1]             $os            = $facts['kernel'] ? {
     'Linux'  => 'linux',
     'Darwin' => 'darwin',
     default  => $facts['kernel'] # lint:ignore:parameter_documentation broken
   },
-  String[1]        $arch          = $facts['os']['hardware'] ? {
+  String[1]             $arch          = $facts['os']['hardware'] ? {
     undef     => 'amd64', # Assume amd64 if os.hardware is missing.
     'aarch64' => 'arm64',
     'armv7l'  => 'armv6l',
@@ -34,15 +38,24 @@ class golang (
     'x86_64'  => 'amd64',
     default   => $facts['os']['hardware'], # lint:ignore:parameter_documentation broken
   },
-  String[1]        $source        = "${source_prefix}/go${version}.${os}-${arch}.tar.gz",
+  String[1]             $source        = "${source_prefix}/go${version}.${os}-${arch}.tar.gz",
 ) {
   $archive_path = '/tmp/puppet-golang.tar.gz'
+
+  $file_ensure = $ensure ? {
+    'present' => file,
+    default   => absent,
+  }
+  $link_ensure = $ensure ? {
+    'present' => link,
+    default   => absent,
+  }
 
   include archive
 
   # Used to ensure that the installation is updated when $source changes.
   file { '/usr/local/share/go-SOURCE':
-    ensure  => file,
+    ensure  => $file_ensure,
     owner   => 0,
     group   => 0, # group might be called root or wheel
     mode    => '0644',
@@ -52,20 +65,27 @@ class golang (
       ${source}
       | EOF
     # lint:endignore
-    notify  => Exec['dp/golang refresh go installation'],
   }
 
-  # If the /usr/local/go directory exists, archive won't update it.
-  exec { 'dp/golang refresh go installation':
-    command     => 'rm -rf /usr/local/go',
-    path        => ['/usr/local/bin', '/usr/bin', '/bin'],
-    user        => 'root',
-    refreshonly => true,
-    notify      => Archive[$archive_path],
+  if $ensure == present {
+    # If the /usr/local/go directory exists, archive won't update it.
+    exec { 'dp/golang refresh go installation':
+      command     => 'rm -rf /usr/local/go',
+      path        => ['/usr/local/bin', '/usr/bin', '/bin'],
+      user        => 'root',
+      refreshonly => true,
+      subscribe   => File['/usr/local/share/go-SOURCE'],
+      notify      => Archive[$archive_path],
+    }
+  } else {
+    file { '/usr/local/go':
+      ensure => absent,
+      force  => true,
+    }
   }
 
   archive { $archive_path:
-    ensure       => present,
+    ensure       => $ensure,
     extract      => true,
     extract_path => '/usr/local',
     source       => $source,
@@ -75,7 +95,7 @@ class golang (
 
   $link_binaries.each |$binary| {
     file { "/usr/local/bin/${binary}":
-      ensure  => link,
+      ensure  => $link_ensure,
       target  => "/usr/local/go/bin/${binary}",
       require => Archive[$archive_path],
     }
